# docs/swagger.yaml
basePath: /api
definitions:
...
  core.JSONB:
    additionalProperties: true
    type: object
...
  database.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  product.Product:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/database.DeletedAt'
      description:
        type: string
      id:
        type: string
      images:
        items:
          type: string
        type: array
      livemode:
        type: boolean
      metadata:
        $ref: '#/definitions/core.JSONB'
      name:
        type: string
      object:
        type: string
      package_dimensions:
        format: uuid
        type: string
      shippable:
        type: boolean
      statement_descriptor:
        type: string
      tax_code:
        format: uuid
        type: string
      unit_label:
        type: string
      updated:
        type: integer
      updated_at:
        type: string
      url:
        type: string
    type: object
  promotion.PromotionCode:
    properties:
      active:
        type: boolean
      code:
        type: string
      coupon:
        format: uuid
        type: string
      created_at:
        type: string
      customer:
        format: uuid
        type: string
      deleted_at:
        $ref: '#/definitions/database.DeletedAt'
      expires_at:
        type: integer
      first_time_transaction:
        type: boolean
      id:
        type: string
      livemode:
        type: boolean
      max_redemptions:
        type: integer
      metadata:
        $ref: '#/definitions/core.JSONB'
      minimum_amount:
        type: integer
      minimum_amount_currency:
        type: string
      object:
        type: string
      times_redeemed:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Swagger Example API
  version: "2.0"
paths:
...
  /v1/products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Product by given ID.
      parameters:
      - description: Product ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "204":
          description: ok
          schema:
            type: string
      summary: delete Product by given ID
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Get Product by given ID or 404 error.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.Product'
      summary: get Product by given ID or 404 error.
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Update Product.
      parameters:
      - description: Product ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.Product'
      summary: update Product
      tags:
      - Product
...
  /v1/promotion_codes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete PromotionCode by given ID.
      parameters:
      - description: PromotionCode ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "204":
          description: ok
          schema:
            type: string
      summary: delete PromotionCode by given ID
      tags:
      - PromotionCode
    get:
      consumes:
      - application/json
      description: Get PromotionCode by given ID or 404 error.
      parameters:
      - description: PromotionCode ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/promotion.PromotionCode'
      summary: get PromotionCode by given ID or 404 error.
      tags:
      - PromotionCode
    post:
      consumes:
      - application/json
      description: Update PromotionCode.
      parameters:
      - description: PromotionCode ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/promotion.PromotionCode'
      summary: update PromotionCode
      tags:
      - PromotionCode
...
schemes:
- http
swagger: "2.0"